# docs/api/openapi.json
{
  "components": {
    "headers": {
      "IdempotentReplay": {
        "description": "Present (value `true`) when this response was replayed from a stored `Idempotency-Key`.",
        "schema": {
          "const": "true",
          "type": "string"
        }
      },
      "RetryAfter": {
        "description": "Seconds to wait before retrying.",
        "schema": {
          "minimum": 0,
          "type": "integer"
        }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "description": "Any string up to 200 characters, unique per request you intend once (a UUID is fine). Honoured on POST for 24 hours per tenant: a replay with the same key AND the same method, URL and body returns the original response — same status, same body — with `Idempotent-Replay: true`. The same key with a different request is a 422 `VALIDATION_FAILED`. Responses ≥ 500 are not stored, so a failed request may be retried with the same key.",
        "in": "header",
        "name": "Idempotency-Key",
        "required": false,
        "schema": {
          "maxLength": 200,
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Malformed JSON, unsupported media type, or a body over 256 KiB (Fastify `FST_ERR_*` code in the envelope)."
      },
      "EngineUnavailable": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`ENGINE_UNAVAILABLE` — a dependency is not configured or is temporarily failing; retry after `Retry-After` when present.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        }
      },
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`FORBIDDEN` — the key lacks the scope, the request IP is outside the key's allow-list (E-70), a public key was used from an Origin outside its domain list or for anything but `lead_callback`, or the account is suspended/uninstalled."
      },
      "Internal": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`INTERNAL` — unexpected error; quote `request_id` to support. Nothing internal is disclosed."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`NOT_FOUND` — the id does not exist, or belongs to another tenant (row-level security makes those indistinguishable)."
      },
      "RateLimited": {
        "content": {
          "application/json": {
            "schema": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/RateLimitError"
                },
                {
                  "$ref": "#/components/schemas/Error"
                }
              ]
            }
          }
        },
        "description": "`RATE_LIMITED`. The per-minute limiter answers with the flat `RateLimitError` body; the per-key daily intent cap (E-70) answers with the standard `Error` envelope and `Retry-After: 3600`.",
        "headers": {
          "Retry-After": {
            "$ref": "#/components/headers/RetryAfter"
          }
        }
      },
      "Unauthenticated": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`UNAUTHENTICATED` — missing, malformed, unknown or revoked API key. Keys look like `nrd_live_…`, `nrd_test_…` or `nrd_pk_…`."
      },
      "ValidationFailed": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "`VALIDATION_FAILED` — the body did not match the schema (`details` lists `{path, message}` issues), a phone number could not be parsed, a timestamp is in the future, or an `Idempotency-Key` was reused with a different request."
      }
    },
    "schemas": {
...
      "DirectionUsage": {
        "description": "Usage for one direction (outbound outcomes or inbound minutes). Keys are camelCase as returned.",
        "properties": {
          "extra": {
            "description": "Units charged beyond the allowance.",
            "type": "integer"
          },
          "extraAmountMinor": {
            "description": "Charged amount for the extra units, in minor currency units (paise/cents).",
            "type": "integer"
          },
          "feeMinor": {
            "description": "Platform fee for the period, minor units.",
            "type": "integer"
          },
          "included": {
            "description": "Units included in the plan for the period (outcomes for outbound, minutes for inbound).",
            "type": "integer"
          },
          "plan": {
            "description": "Plan code, or null when no plan is attached for this direction.",
            "type": [
              "string",
              "null"
            ]
          },
          "unitMinor": {
            "description": "Price per extra unit, minor units.",
            "type": "integer"
          },
          "used": {
            "description": "Units consumed this period.",
            "type": "integer"
          }
        },
        "required": [
          "plan",
          "included",
          "used",
          "extra",
          "extraAmountMinor",
          "unitMinor",
          "feeMinor"
        ],
        "type": "object"
      },
      "Dispute": {
        "properties": {
          "id": {
            "description": "26-character ULID with a type prefix",
            "type": "string"
          },
          "opened_at": {
            "format": "date-time",
            "type": "string"
          },
          "outcome_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "resolution": {
            "description": "Staff note when accepted or rejected.",
            "type": [
              "string",
              "null"
            ]
          },
          "resolved_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "enum": [
              "open",
              "accepted",
              "rejected"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "outcome_id",
          "status",
          "reason",
          "resolution",
          "opened_at",
          "resolved_at"
        ],
        "type": "object"
      },
      "Error": {
        "description": "The error envelope every non-2xx response uses, except the per-minute rate limiter (see RateLimitError).",
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Machine-readable code. Status by code: GATED → 409, ENGINE_UNAVAILABLE → 503, IDEMPOTENT_REPLAY → 200, SIGNATURE_INVALID → 401, VALIDATION_FAILED → 422, UNAUTHENTICATED → 401, FORBIDDEN → 403, RATE_LIMITED → 429, NOT_FOUND → 404, DISPATCH_UNCERTAIN → 500, INTERNAL → 500. Fastify's own client errors (malformed JSON, unsupported media type, body over 256 KiB) use Fastify's `FST_ERR_*` codes with the same envelope.",
                "enum": [
                  "GATED",
                  "ENGINE_UNAVAILABLE",
                  "IDEMPOTENT_REPLAY",
                  "SIGNATURE_INVALID",
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "RATE_LIMITED",
                  "NOT_FOUND",
                  "DISPATCH_UNCERTAIN",
                  "INTERNAL"
                ],
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "description": "Optional. For VALIDATION_FAILED from the schema: an array of `{path, message}` issues. Otherwise the error context (flat key/values), when any.",
                "type": "object"
              },
              "message": {
                "description": "Human-readable, safe to show to an operator; never contains a phone number.",
                "type": "string"
              },
              "request_id": {
                "description": "Echo of the request id (`x-cloud-trace-context` when present). Quote it in support requests.",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "RateLimitError": {
        "description": "The flat body of the per-minute limiter (@fastify/rate-limit). The per-key DAILY intent cap answers 429 with the standard Error envelope instead.",
        "properties": {
          "code": {
            "const": "RATE_LIMITED",
            "type": "string"
          },
          "error": {
            "const": "Too Many Requests",
            "type": "string"
          },
          "message": {
            "description": "e.g. \"rate limit exceeded, retry in 12s\"",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "statusCode": {
            "const": 429,
            "type": "integer"
          }
        },
        "required": [
          "statusCode",
          "code",
          "error",
          "message",
          "request_id"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "publicSiteKey": {
        "bearerFormat": "nrd_pk_…",
        "description": "Public site key for browsers: `POST /v1/intents` with `use_case: lead_callback` only, from an allow-listed Origin, 10 requests/minute per IP.",
        "scheme": "bearer",
        "type": "http"
      },
      "secretKey": {
        "bearerFormat": "nrd_live_… / nrd_test_…",
        "description": "Secret API key. The scope names listed on each operation are what the key must carry.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Naaradh",
      "url": "https://naaradh.com"
    },
...
    "summary": "Two-way AI voice agent for commerce — outbound call intents and the inbound support line.",
    "title": "Naaradh API",
    "version": "0.0.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/v1/billing": {
      "get": {
        "description": "What the billing page shows: the calendar-month period, plan and allowance per direction, units used and charged beyond the allowance, credits from accepted disputes, the state of usage postings at the provider, and the subscription as Naaradh last fetched it. Amounts are integers in minor units (paise/cents) with the currency alongside.\n\nScope: `billing:read`.",
        "operationId": "getBilling",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "billing_provider": {
                      "enum": [
                        "shopify",
                        "razorpay",
                        "stripe",
                        "manual"
                      ],
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "billing_status": {
                      "enum": [
                        "none",
                        "active",
                        "frozen",
                        "capped",
                        "cancelled"
                      ],
                      "type": "string"
                    },
                    "credits_minor": {
                      "description": "Credits applied this period (positive number), minor units.",
                      "type": "integer"
                    },
                    "currency": {
                      "enum": [
                        "INR",
                        "USD"
                      ],
                      "type": "string"
                    },
                    "grace_until": {
                      "description": "End of the 3-day grace period after a declined payment (E-50).",
                      "format": "date-time",
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "inbound": {
                      "$ref": "#/components/schemas/DirectionUsage"
                    },
                    "outbound": {
                      "$ref": "#/components/schemas/DirectionUsage"
                    },
                    "period": {
                      "description": "`YYYY-MM` (UTC).",
                      "type": "string"
                    },
                    "postings": {
                      "additionalProperties": {
                        "type": "integer"
                      },
                      "description": "Count of usage postings by status this period. Keys: pending, posted, capped, failed, skipped.",
                      "type": "object"
                    },
                    "subscription": {
                      "description": "The most recent subscription row, or null.",
                      "properties": {
                        "cappedAmountMinor": {
                          "description": "Shopify capped amount (E-61), minor units.",
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "currentPeriodEnd": {
                          "description": "Same value as `current_period_end` (both keys are returned).",
                          "format": "date-time",
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "current_period_end": {
                          "format": "date-time",
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "provider": {
                          "enum": [
                            "shopify",
                            "razorpay",
                            "stripe",
                            "manual"
                          ],
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "pending",
                            "active",
                            "frozen",
                            "cancelled",
                            "declined",
                            "expired"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "provider",
                        "status",
                        "currentPeriodEnd",
                        "cappedAmountMinor",
                        "current_period_end"
                      ],
                      "type": [
                        "object",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "period",
                    "currency",
                    "billing_status",
                    "billing_provider",
                    "grace_until",
                    "outbound",
                    "inbound",
                    "credits_minor",
                    "postings",
                    "subscription"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Usage summary."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "billing:read"
            ]
          }
        ],
        "summary": "Plan, allowance and usage for the current period",
        "tags": [
          "Billing"
        ]
      }
    },
    "/v1/billing/razorpay/subscribe": {
      "post": {
        "description": "Creates a Razorpay subscription for the chosen plan combination and returns the mandate authorisation URL; the subscription becomes active when Razorpay's webhook is confirmed by a re-fetch. Outbound plan codes: `starter`, `growth`, `scale`, `enterprise`. Support-line plan codes: `inbound_starter`, `inbound_growth`, `inbound_scale`. At least one of the two is required. 403 when the account is a Shopify install (billed through Shopify only); 422 when the combination is not offered; 503 when Razorpay is not configured or unavailable.\n\nScope: `billing:write`.",
        "operationId": "subscribeRazorpay",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "inbound_plan_code": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null
                  },
                  "plan_code": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "default": null
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "authorize_url": {
                      "format": "uri",
                      "type": "string"
                    },
                    "status": {
                      "const": "pending",
                      "type": "string"
                    },
                    "subscription_id": {
                      "description": "26-character ULID with a type prefix",
                      "type": "string"
                    }
                  },
                  "required": [
                    "subscription_id",
                    "status",
                    "authorize_url"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Created; send the merchant to `authorize_url`.",
            "headers": {
              "Idempotent-Replay": {
                "$ref": "#/components/headers/IdempotentReplay"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          },
          "503": {
            "$ref": "#/components/responses/EngineUnavailable"
          }
        },
        "security": [
          {
            "secretKey": [
              "billing:write"
            ]
          }
        ],
        "summary": "Start a Razorpay subscription (direct INR merchants)",
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/v1/disputes": {
      "get": {
        "description": "The last 200 disputes for your account, newest first, with their resolution when decided.\n\nScope: `billing:read`.",
        "operationId": "listDisputes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "items": {
                        "$ref": "#/components/schemas/Dispute"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Disputes."
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "billing:read"
            ]
          }
        ],
        "summary": "List disputes",
        "tags": [
          "Billing"
        ]
      }
    },
...
    "/v1/outcomes/{id}/disputes": {
      "post": {
        "description": "Opens a dispute on a BILLED outcome within 7 days of billing (E-62). Staff review the recording, transcript and extraction; an accepted dispute becomes a credit on the next period (the ledger is append-only, so the original charge is never edited). 422 when the outcome was not billed, was within the plan allowance, or the window has passed.\n\nScope: `billing:write`.",
        "operationId": "openOutcomeDispute",
        "parameters": [
          {
            "description": "Outcome id (`out_…`) from the intent's `outcome.outcome_id` or the `outcome.final` webhook.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "reason": {
                    "description": "Why you believe the outcome should not have been billed (10–2000 characters).",
                    "maxLength": 2000,
                    "minLength": 10,
                    "type": "string"
                  }
                },
                "required": [
                  "reason"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "dispute_id": {
                      "description": "26-character ULID with a type prefix",
                      "type": "string"
                    },
                    "status": {
                      "const": "open",
                      "type": "string"
                    }
                  },
                  "required": [
                    "dispute_id",
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Opened.",
            "headers": {
              "Idempotent-Replay": {
                "$ref": "#/components/headers/IdempotentReplay"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "security": [
          {
            "secretKey": [
              "billing:write"
            ]
          }
        ],
        "summary": "Dispute a billed outcome",
        "tags": [
          "Billing"
        ]
      }
    },
...
  },
  "security": [
    {
      "secretKey": []
    }
  ],
  "servers": [
    {
      "description": "Production",
      "url": "https://api.naaradh.com"
    },
    {
      "description": "Local development (apps/api default PORT)",
      "url": "http://localhost:3001"
    }
  ],
...
}